fixing overlay build size
only importing abstract classes in the overlay preload script
