+++ app-back/Services/Storage/abstract.ts
@@ -1,4 +1,4 @@
-import { Service } from "../abstract";
+import { Service } from "../service";
 
 export type Game = {
     id: string;
