+++ app-back/preloadOverlay.ts
@@ -1,15 +1,32 @@
-import { availableServices } from './Services/available';
+// import { abstractServices } from './Services/available';
+import AbstractScanner from "./Services/Scanner/abstract";
+import AbstractStorage from "./Services/Storage/abstract";
+import AbstractEmulatorManager from "./Services/Emulator/abstractManager";
+import AbstractMetadata from "./Services/Metadata/abstract";
+import AbstractModManager from "./Services/Mods/abstractManager";
+import AbstractAccount from "./Services/Account/abstract";
+import AbstractController from "./Services/Controller/abstract";
 const { contextBridge, ipcRenderer } = require('electron');
+
+export const abstractServices = {
+    Account: AbstractAccount,
+    Emulator: AbstractEmulatorManager,
+    Metadata: AbstractMetadata,
+    Mods: AbstractModManager,
+    Scanner: AbstractScanner,
+    Storage: AbstractStorage,
+    Controller: AbstractController,
+}
 
 const track = (path, data) => ipcRenderer.invoke('subway-track', { path, data });
 
 const servicesSubway = Object.fromEntries(
-    Object.entries(availableServices)
+    Object.entries(abstractServices)
     .map(
         (entry) => [
             entry[0],
             Object.fromEntries(
-                Object.getOwnPropertyNames(entry[1].abstract.prototype)
+                Object.getOwnPropertyNames(entry[1].prototype)
                 .filter((p) => !["constructor", "setServices", "clean"].includes(p))
                 .map((p) => [
                     p,
